complete user routers add exception handler add exceptions

+++ Ego_users-microservice/src/application/user/use_cases/avatar/delete_avatar.py
@@ -28,10 +28,13 @@ class DeleteAvatar(BaseUseCase):
         self._mapper = mapper
         self._uow = uow
 
-    async def __call__(self, data: DeleteAvatarData) -> dto.DeletedAvatarDTO:
+    async def __call__(self, data: DeleteAvatarData) -> dto.DeletedAvatarDTO | None:
         avatar = await self._uow.avatar_repo.get_avatar_by_user_id(
             avatar_user_id=AvatarUserId(value=data.avatar_user_id)
         )
+
+        if not avatar:
+            return avatar  # type: ignore
 
         avatar.delete()
         await self._uow.avatar_repo.delete_avatar(avatar_id=avatar.avatar_id)
